test: dedupe award-history test fixtures into conftest and tests/utils
